test: split whatwg-url-inspect test
